+++ changelog.md
@@ -1,4 +1,8 @@
 # ACEmulator Change Log
+### 2017-10-30
+[Mogwai]
+* implemented updating of weenies in ace-world necessary for crowd-sourced content creation
+* added database unit tests to AppVeyor build
 
 ### 2017-11-01
 [Ripley]